requirements.txt
click==8.1.7
rich==13.7.1
PyYAML==6.0.2
